requirements.txt
psycopg2-binary==2.9.9
boto3==1.34.34
requests==2.31.0
